The Multi-Personality Effect: Different Makes Perfect
We are often told about the ways a team can become more productive. From processes, methodologies and ensuring the members are 
multi-disciplinary, there seems to be something often disregarded: the multi-personality effect.

Have you ever been assigned to a team where you feel like a stranger? Where you have to be
careful about how you talk to people because you don't know how they'll take it?
Where you don't know if your work methodologies are the norm? And if they're not, you don't know what the norm is.

Two years ago, I joined a new team with a special "feature": we were all completely different.
It became obvious that we had a lot to learn about each other and also from each other.

The challenge ahead was how to define the best strategies to work together and be a productive team.

In this presentation, you will be able to see how the multi-personality effect was actually a good thing. We found a way to make 
it work and even better... we became a highly productive team, while having fun at the same time!