A Team - Turn a bunch of developers into a team of A players ============================================================
* Speaker   : Ricardo Castelhano
* Available : first day, second day, third day
* Length    : 45 minutes + QA
* Language  : Portuguese (might be in English if by attendees request

Description
-----------

So, you just inherited a supposed team of developers, but you can easily find out that they are just a bunch of developers. They are totally dysfunctional as a team. But individually, each one is a valid contributor. You don't want to give up of them. Will you be able to be the much needed glue to join them? And the trigger to send them to the A class of contributors?

Speaker Bio
-----------

I consider myself a creative and curious mind. I made my journey from Mathematics into Creative Computing, then into Web Development and lately to Management. I consider myself a people person, a good communicator, and a good listener. I prefer to play as a team, with a special taste for motivating and helping those around me to achieve better results.

Links
-----

* Blog: http://ricardocastelhano.net
* Company: RUPEAL (InvoiceXpress, ClanHR, KWAN)
* GitHub: https://github.com/RicCastelhano
* Photo: http://ricardocastelhano.net/wp-content/uploads/2019/11/EU.jpg

Extra Information
-----------------

A shortlist of some past presentations I did.
http://ricardocastelhano.net/talks/
Some have links for the slidedeck or the session recordings.